splitting up gameobject
