+++ UnityMcpBridge/Editor/Tools/ManageGameObject.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 152279796ecc3d6e9903f8b6985d6cdb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
