+++ UnityMcpBridge/Editor/Tools/ManageGameObject/ComponentManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: acded639db7b1abe8ab5ab9fbc3bb527
