+++ UnityMcpBridge/Editor/Tools/ManageGameObject/GameObjectCreator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7f304ae81c15e5869b220134b740a7ce
