+++ UnityMcpBridge/Editor/Tools/ManageGameObject/GameObjectDeleter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8b91f624d689528dab866ea114ee3862
