+++ UnityMcpBridge/Editor/Tools/ManageGameObject/GameObjectFinder.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0f19b1c1ee48908eda28381ca7d8b48f
