+++ UnityMcpBridge/Editor/Tools/ManageGameObject/GameObjectModifier.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2971f183b8ac59c9e97acd14202d855e
