+++ UnityMcpBridge/Editor/Tools/ManageGameObject/GameObjectSerializer.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cbaeb35220db8c712bc0883b30013609
