+++ UnityMcpBridge/Editor/Tools/ManageGameObject/Models.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b667fa6846e58836b09b9bce0260f34
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
