+++ UnityMcpBridge/Editor/Tools/ManageGameObject/Models/GameObjectModels.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 76603e295d503458bad2c429e7dcc625
