+++ UnityMcpBridge/Editor/Tools/ManageGameObject/PropertyUtils.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a692638d412f6c542baa016b37683228
