+++ UnityMcpBridge/Editor/Tools/ManageGameObject/README.md
@@ -1,0 +1,38 @@
+# ManageGameObject Implementation
+
+This directory contains the implementation details of the `ManageGameObject` tool, which has been refactored from a monolithic class into a more maintainable structure.
+
+## Architecture
+
+The implementation follows a modular approach where functionality is split based on responsibilities:
+
+- `ManageGameObject.cs` (in parent directory): Main entry point that maintains the public API surface and routes commands to the appropriate specialized classes.
+- `GameObjectModels/`: Contains shared data structures used across the implementation.
+- Specialized implementation classes:
+  - `GameObjectCreator.cs`: Handles creation of GameObjects
+  - `GameObjectModifier.cs`: Handles modification of existing GameObjects
+  - `GameObjectDeleter.cs`: Handles deletion of GameObjects
+  - `GameObjectFinder.cs`: Handles finding GameObjects in the scene
+  - `ComponentManager.cs`: Handles component operations (add, remove, modify)
+  - `GameObjectSerializer.cs`: Handles serialization of GameObjects
+  - `PropertyUtils.cs`: Utility functions for setting properties on objects
+
+## Design Principles
+
+1. **Separation of Concerns**: Each class handles a specific set of related functionality.
+2. **Encapsulation**: Implementation details are marked as `internal` to maintain a clean public API.
+3. **Consistency**: All implementation classes follow similar patterns and code style.
+4. **Maintainability**: Smaller files focused on specific tasks make the code easier to understand and maintain.
+
+## Extending the Implementation
+
+When adding new functionality:
+
+1. Determine which existing class should contain the functionality based on responsibility.
+2. If the functionality doesn't fit existing classes, consider creating a new specialized class.
+3. Add appropriate methods to the `ManageGameObject` main class to expose the functionality.
+4. Follow the existing patterns for parameter validation, error handling, and response formatting.
+
+## Usage
+
+The main `ManageGameObject` class is the public entry point and should be used by other parts of the system. The implementation classes in this directory are internal and should not be referenced directly from outside this module.
